cunit

+++ mini_project/3_implementation/src/calculator.c
@@ -10,12 +10,12 @@ int subtract(int operand1, int operand2)
     return operand1 - operand2;
 }
 
-double multiply(int operand1, int operand2)
+double multiply(double operand1, double operand2)
 {
     return operand1 * operand2;
 }
 
-double divide(int operand1, int operand2)
+double divide(double operand1, double operand2)
 {
     if(0 == operand2)
         return 0;
